renames components in kabab case
