+++ client/src/components/byline-metabox/byline-profiles.js
@@ -13,6 +13,8 @@ import Autocomplete from 'react-autocomplete';
 import PropTypes from 'prop-types';
 import { Button } from '@wordpress/components';
 
+// import useDebounce from '../../services/use-debounce';
+
 const SortableItem = SortableElement(({
   count,
   bylineId,
