+++ client/src/components/byline-slot/byline-list/index.jsx
