+++ client/src/index.js
@@ -1,6 +1,6 @@
 // Entry point for all shared JS.
-import initBylineMetaBox from './initBylineMetaBox';
-import initUserLinkMetaBox from './initUserLinkMetaBox';
+import initBylineMetaBox from './init-byline-metabox';
+import initUserLinkMetaBox from './init-user-link-metabox';
 
 import './styles/styles.scss';
