+++ client/src/init-byline-metabox.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import BylineMetaBox from 'components/BylineMetaBox';
+import BylineMetaBox from 'components/byline-metabox';
 
 const initBylineMetaBox = () => {
   ReactDOM.render(
