+++ client/src/init-user-link-metabox.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import UserLinkMetaBox from 'components/UserLinkMetaBox';
+import UserLinkMetaBox from 'components/user-link-metabox';
 
 const initUserLinkMetaBox = () => {
   const userLinkEl = document.getElementById('byline-manager-user-link-root');
